requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpc-split_1662408188328/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964070496/work
idna==3.4
importlib-metadata==5.0.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot @ file:///Users/runner/miniforge3/conda-bld/pydot_1654961384348/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.8.0
tensorflow-metal==0.4.0
termcolor==2.0.1
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
